Setup subscribers list and script

+++ Assets/Scripts/Enemy/Subs/Subscriber.cs
@@ -5,7 +5,7 @@ namespace nickmaltbie.Treachery.Enemy.Subs
     public class Subscriber
     {
         public string name;
-        public int months;
+        public float months;
         public float strength;
     }
 }
